// frontend/src/generated/api/models/CellInfo.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type CellInfo = {
    cellRow: number;
    cellColumn: number;
    cropItem: string;
    variety: string;
    datePlanted: string;
    growthStage: number;
    status: number;
    poorGrowthReason: number;
    dateHarvestPlanted: string;
    memo: string;
};